+++ src/components/GamePlay/ChaseCam.tsx
@@ -1,6 +1,6 @@
 import React, { useRef, useEffect, useState, useContext, } from 'react';
 import { useAuth0 } from '@auth0/auth0-react';
-import { useWebcam } from '../contexts/WebcamChaseProvider';
+import { useWebcam } from '../../contexts/WebcamChaseProvider';
 
 import {
   LocationBasedLocal,
@@ -13,8 +13,8 @@ import {
   TextureLoader,
 } from "./webcam.js"
 
-import SocketContext from '../contexts/Socket/SocketContext';
-import Radar from './Radar/Radar';
+import SocketContext from '../../contexts/Socket/SocketContext';
+import Radar from '../Radar/Radar';
 
 
 const ChaseCam: React.FC = () => {
