+++ src/components/GamePlay/KillProgress.tsx
