+++ src/components/GamePlay/webcam.js
