+++ src/components/Trophies/SingleTrophy.tsx
