+++ src/pages/TrophyRoom.tsx
@@ -5,7 +5,7 @@ import { Main } from '../styles/Main';
 import PhoneLoader from '../components/Loaders/PhoneLoader';
 import InfoPopup from '../components/Popups/InfoPopup';
 
-const SavedTrophies = lazy(() => import('../components/SavedTrophies'));
+const SavedTrophies = lazy(() => import('../components/Trophies/SavedTrophies'));
 
 const TrophyRoom: React.FC = () => {
 
@@ -14,7 +14,7 @@ const TrophyRoom: React.FC = () => {
   return (
     <>
       <Header page='Trophies' />
-      <Main style={{height: '100vh'}}>
+      <Main style={{ height: '100vh' }}>
         <div className='content__body'>
           <div style={{ width: '300px', height: '300px' }}>
             <Suspense fallback={<PhoneLoader />}>
